# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClarkNav API",
        "description": "API documentation for ClarkNav",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bug-reports": {
            "get": {
                "tags": [
                    "Bug Reports"
                ],
                "summary": "Get all bug reports",
                "operationId": "0c69a79bde946dabd8b0e80ce6bb68f5",
                "responses": {
                    "200": {
                        "description": "List of bug reports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "App Crashes When Searching for Routes"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "UI/UX Issue"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "The app crashes when I try to search for a route using public transit."
                                            },
                                            "steps": {
                                                "type": "string",
                                                "example": "1. Open the app\\n2. Search for a route\\n3. App crashes"
                                            },
                                            "expected": {
                                                "type": "string",
                                                "example": "The app should display the search results."
                                            },
                                            "actual": {
                                                "type": "string",
                                                "example": "The app crashes."
                                            },
                                            "device": {
                                                "type": "string",
                                                "example": "iPhone 12, iOS 14.4"
                                            },
                                            "frequency": {
                                                "type": "string",
                                                "example": "Always"
                                            },
                                            "screenshots": {
                                                "type": "string",
                                                "example": "screenshots/example.png",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bug Reports"
                ],
                "summary": "Create a new bug report",
                "operationId": "b554e0b8c6b4d789c8c64e088f353e8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "description",
                                    "steps",
                                    "expected",
                                    "actual",
                                    "device",
                                    "frequency"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "App Crashes When Searching for Routes"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "UI/UX Issue"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The app crashes when I try to search for a route using public transit."
                                    },
                                    "steps": {
                                        "type": "string",
                                        "example": "1. Open the app\\n2. Search for a route\\n3. App crashes"
                                    },
                                    "expected": {
                                        "type": "string",
                                        "example": "The app should display the search results."
                                    },
                                    "actual": {
                                        "type": "string",
                                        "example": "The app crashes."
                                    },
                                    "device": {
                                        "type": "string",
                                        "example": "iPhone 12, iOS 14.4"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "Always"
                                    },
                                    "screenshots": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bug report created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "App Crashes When Searching for Routes"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "UI/UX Issue"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "The app crashes when I try to search for a route using public transit."
                                        },
                                        "steps": {
                                            "type": "string",
                                            "example": "1. Open the app\\n2. Search for a route\\n3. App crashes"
                                        },
                                        "expected": {
                                            "type": "string",
                                            "example": "The app should display the search results."
                                        },
                                        "actual": {
                                            "type": "string",
                                            "example": "The app crashes."
                                        },
                                        "device": {
                                            "type": "string",
                                            "example": "iPhone 12, iOS 14.4"
                                        },
                                        "frequency": {
                                            "type": "string",
                                            "example": "Always"
                                        },
                                        "screenshots": {
                                            "type": "string",
                                            "example": "screenshots/example.png",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}